requirements.txt
torch==1.8.1
torchvision~=0.14.1
tensorboard==2.7.0
opencv-python~=4.7.0.72
numpy~=1.23.5
matplotlib~=3.7.1
pillow~=9.4.0
tqdm~=4.65.0